+++ common/views/components/BaseTabs/README.md
